backendBeninHeart/requirements.txt
# Django and DRF
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
drf-spectacular>=0.29.0

# Environment variables
python-decouple>=3.8

# Database
psycopg2-binary>=2.9  # PostgreSQL

# Image handling
Pillow>=10.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# HTTP Requests (for Cloudflare Turnstile verification)
requests>=2.31.0

# Production server
gunicorn>=21.2.0

# Redis (caching and sessions)
redis>=5.0.0
django-redis>=5.3.0

# Celery (async tasks)
celery>=5.3.0
flower>=2.0.0  # Celery monitoring

# Development
django-debug-toolbar>=4.2
django-extensions>=3.2
# Filtering
django-filter>=23.0